+++ staypresent/__init__.py
@@ -2,5 +2,5 @@ from .runner import run
 from .pinger import ping, cron, active_cron_handles
 from . import web
 
-__version__ = "1.5.7"
+__version__ = "1.5.8"
 __all__ = ["run", "web", "ping", "cron", "active_cron_handles", "__version__"]